Fixing the issue : Vertical scrollbar isn't shown when verticalScrollSpec is used even if scrollbarEnabled is true by default. The issue is with NestedScrollView when used programatically doesn't initialize scrollBars. NestedScrollView when used in xml initializes scrollbars.
